requirements.txt
pandas
click
requests
python-dotenv
pymongo
typer